feat: integrate the solx compiler
