+++ compiler_tester/Cargo.toml
@@ -52,6 +52,8 @@ era-compiler-llvm-context = { git = "https://github.com/matter-labs/era-compiler
 era-compiler-solidity = { git = "https://github.com/matter-labs/era-compiler-solidity", branch = "main" }
 era-solc = { git = "https://github.com/matter-labs/era-compiler-solidity", branch = "main" }
 era-compiler-vyper = { git = "https://github.com/matter-labs/era-compiler-vyper", branch = "main" }
+solx = { git = "https://github.com/matter-labs/solx", branch = "main" }
+solx-solc = { git = "https://github.com/matter-labs/solx", branch = "main" }
 
 solidity-adapter = { path = "../solidity_adapter" }
 benchmark-analyzer = { path = "../benchmark_analyzer" }
