+++ compiler_tester/src/compilers/solidity/solc/compiler/standard_json/input/settings/debug.rs
