+++ compiler_tester/src/lib.rs
@@ -7,6 +7,7 @@
 #![allow(clippy::too_many_arguments)]
 #![allow(clippy::type_complexity)]
 
+pub(crate) mod benchmark_format;
 pub(crate) mod compilers;
 pub(crate) mod directories;
 pub(crate) mod environment;
@@ -26,14 +27,15 @@ use itertools::Itertools;
 use rayon::iter::IntoParallelIterator;
 use rayon::iter::ParallelIterator;
 
+pub use crate::benchmark_format::BenchmarkFormat;
 pub use crate::compilers::eravm::EraVMCompiler;
 pub use crate::compilers::llvm::LLVMCompiler;
 pub use crate::compilers::mode::llvm_options::LLVMOptions;
 pub use crate::compilers::mode::Mode;
-pub use crate::compilers::solidity::mode::Mode as SolidityMode;
-pub use crate::compilers::solidity::upstream::solc::standard_json::input::language::Language as SolcStandardJsonInputLanguage;
-pub use crate::compilers::solidity::upstream::SolidityCompiler as SolidityUpstreamCompiler;
-pub use crate::compilers::solidity::SolidityCompiler;
+pub use crate::compilers::solidity::solc::compiler::standard_json::input::language::Language as SolcStandardJsonInputLanguage;
+pub use crate::compilers::solidity::solc::SolidityCompiler as SolcCompiler;
+pub use crate::compilers::solidity::zksolc::mode::Mode as ZksolcMode;
+pub use crate::compilers::solidity::zksolc::SolidityCompiler as ZksolcCompiler;
 pub use crate::compilers::vyper::VyperCompiler;
 pub use crate::compilers::yul::YulCompiler;
 pub use crate::compilers::Compiler;
@@ -55,6 +57,9 @@ pub use crate::workflow::Workflow;
 
 /// The debug directory path.
 pub const DEBUG_DIRECTORY: &str = "./debug/";
+
+/// The default benchmark context.
+pub const ARGUMENT_BENCHMARK_CONTEXT: &str = "benchmark-context";
 
 ///
 /// The compiler test generic representation.
@@ -231,12 +236,13 @@ impl CompilerTester {
         target: era_compiler_common::Target,
         toolchain: Toolchain,
     ) -> anyhow::Result<Vec<Test>> {
-        let solidity_compiler = Arc::new(SolidityCompiler::new());
-        let solidity_upstream_compiler = Arc::new(SolidityUpstreamCompiler::new(
+        let zksolc_compiler = Arc::new(ZksolcCompiler::new());
+        let solx_compiler = Arc::new(SolxCompiler::new());
+        let solc_compiler = Arc::new(SolcCompiler::new(
             SolcStandardJsonInputLanguage::Solidity,
             toolchain,
         ));
-        let solidity_upstream_yul_compiler = Arc::new(SolidityUpstreamCompiler::new(
+        let solc_yul_compiler = Arc::new(SolcCompiler::new(
             SolcStandardJsonInputLanguage::Yul,
             toolchain,
         ));
@@ -251,26 +257,27 @@ impl CompilerTester {
             target,
             PathBuf::from(Self::SOLIDITY_SIMPLE.replace("/", std::path::MAIN_SEPARATOR_STR)),
             era_compiler_common::EXTENSION_SOLIDITY,
-            match toolchain {
-                Toolchain::IrLLVM => solidity_compiler.clone(),
-                Toolchain::Solc | Toolchain::SolcLLVM => solidity_upstream_compiler.clone(),
+            match (target, toolchain) {
+                (era_compiler_common::Target::EraVM, Toolchain::IrLLVM) => zksolc_compiler.clone(),
+                (era_compiler_common::Target::EVM, Toolchain::IrLLVM) => solx_compiler.clone(),
+                (_, Toolchain::Solc | Toolchain::SolcLLVM) => solc_compiler.clone(),
             },
         )?);
-        if let era_compiler_common::Target::EraVM = target {
-            tests.extend(self.directory::<MatterLabsDirectory>(
-                target,
-                PathBuf::from(Self::VYPER_SIMPLE.replace("/", std::path::MAIN_SEPARATOR_STR)),
-                era_compiler_common::EXTENSION_VYPER,
-                vyper_compiler.clone(),
-            )?);
-        }
+        // if let era_compiler_common::Target::EraVM = target {
+        //     tests.extend(self.directory::<MatterLabsDirectory>(
+        //         target,
+        //         PathBuf::from(Self::VYPER_SIMPLE.replace("/", std::path::MAIN_SEPARATOR_STR)),
+        //         era_compiler_common::EXTENSION_VYPER,
+        //         vyper_compiler.clone(),
+        //     )?);
+        // }
         tests.extend(self.directory::<MatterLabsDirectory>(
             target,
             PathBuf::from(Self::YUL_SIMPLE.replace("/", std::path::MAIN_SEPARATOR_STR)),
             era_compiler_common::EXTENSION_YUL,
             match toolchain {
                 Toolchain::IrLLVM => yul_compiler.clone(),
-                Toolchain::Solc | Toolchain::SolcLLVM => solidity_upstream_yul_compiler.clone(),
+                Toolchain::Solc | Toolchain::SolcLLVM => solc_yul_compiler.clone(),
             },
         )?);
         tests.extend(self.directory::<MatterLabsDirectory>(
@@ -279,30 +286,33 @@ impl CompilerTester {
             era_compiler_common::EXTENSION_LLVM_SOURCE,
             llvm_compiler,
         )?);
-        tests.extend(self.directory::<MatterLabsDirectory>(
-            target,
-            PathBuf::from(Self::ERAVM_SIMPLE.replace("/", std::path::MAIN_SEPARATOR_STR)),
-            era_compiler_common::EXTENSION_ERAVM_ASSEMBLY,
-            eravm_compiler,
-        )?);
+        if let era_compiler_common::Target::EraVM = target {
+            tests.extend(self.directory::<MatterLabsDirectory>(
+                target,
+                PathBuf::from(Self::ERAVM_SIMPLE.replace("/", std::path::MAIN_SEPARATOR_STR)),
+                era_compiler_common::EXTENSION_ERAVM_ASSEMBLY,
+                eravm_compiler,
+            )?);
+        }
 
         tests.extend(self.directory::<MatterLabsDirectory>(
             target,
             PathBuf::from(Self::SOLIDITY_COMPLEX.replace("/", std::path::MAIN_SEPARATOR_STR)),
             era_compiler_common::EXTENSION_JSON,
-            match toolchain {
-                Toolchain::IrLLVM => solidity_compiler.clone(),
-                Toolchain::Solc | Toolchain::SolcLLVM => solidity_upstream_compiler.clone(),
+            match (target, toolchain) {
+                (era_compiler_common::Target::EraVM, Toolchain::IrLLVM) => zksolc_compiler.clone(),
+                (era_compiler_common::Target::EVM, Toolchain::IrLLVM) => solx_compiler.clone(),
+                (_, Toolchain::Solc | Toolchain::SolcLLVM) => solc_compiler.clone(),
             },
         )?);
-        if let era_compiler_common::Target::EraVM = target {
-            tests.extend(self.directory::<MatterLabsDirectory>(
-                target,
-                PathBuf::from(Self::VYPER_COMPLEX.replace("/", std::path::MAIN_SEPARATOR_STR)),
-                era_compiler_common::EXTENSION_JSON,
-                vyper_compiler.clone(),
-            )?);
-        }
+        // if let era_compiler_common::Target::EraVM = target {
+        //     tests.extend(self.directory::<MatterLabsDirectory>(
+        //         target,
+        //         PathBuf::from(Self::VYPER_COMPLEX.replace("/", std::path::MAIN_SEPARATOR_STR)),
+        //         era_compiler_common::EXTENSION_JSON,
+        //         vyper_compiler.clone(),
+        //     )?);
+        // }
 
         tests.extend(
             self.directory::<EthereumDirectory>(
@@ -315,20 +325,23 @@ impl CompilerTester {
                     .replace("/", std::path::MAIN_SEPARATOR_STR),
                 ),
                 era_compiler_common::EXTENSION_SOLIDITY,
-                match toolchain {
-                    Toolchain::IrLLVM => solidity_compiler.clone(),
-                    Toolchain::Solc | Toolchain::SolcLLVM => solidity_upstream_compiler.clone(),
+                match (target, toolchain) {
+                    (era_compiler_common::Target::EraVM, Toolchain::IrLLVM) => {
+                        zksolc_compiler.clone()
+                    }
+                    (era_compiler_common::Target::EVM, Toolchain::IrLLVM) => solx_compiler.clone(),
+                    (_, Toolchain::Solc | Toolchain::SolcLLVM) => solc_compiler.clone(),
                 },
             )?,
         );
-        if let era_compiler_common::Target::EraVM = target {
-            tests.extend(self.directory::<EthereumDirectory>(
-                target,
-                PathBuf::from(Self::VYPER_ETHEREUM.replace("/", std::path::MAIN_SEPARATOR_STR)),
-                era_compiler_common::EXTENSION_VYPER,
-                vyper_compiler,
-            )?);
-        }
+        // if let era_compiler_common::Target::EraVM = target {
+        //     tests.extend(self.directory::<EthereumDirectory>(
+        //         target,
+        //         PathBuf::from(Self::VYPER_ETHEREUM.replace("/", std::path::MAIN_SEPARATOR_STR)),
+        //         era_compiler_common::EXTENSION_VYPER,
+        //         vyper_compiler,
+        //     )?);
+        // }
 
         Ok(tests)
     }
