+++ compiler_tester/src/toolchain.rs
@@ -7,8 +7,10 @@
 ///
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, serde::Deserialize)]
 pub enum Toolchain {
-    /// The LLVM toolchain processing IRs from source level compilers
+    /// The default LLVM-based compiler: `zksolc` for EraVM, `solx` for EVM.
     IrLLVM,
+    /// The `zksolc` compiler.
+    ZkSolc,
     /// The upstream `solc` compiler.
     Solc,
     /// The forked `solc` compiler with MLIR.
@@ -21,6 +23,7 @@ impl std::str::FromStr for Toolchain {
     fn from_str(string: &str) -> Result<Self, Self::Err> {
         match string {
             "ir-llvm" => Ok(Self::IrLLVM),
+            "zksolc" => Ok(Self::ZkSolc),
             "solc" => Ok(Self::Solc),
             "solc-llvm" => Ok(Self::SolcLLVM),
             string => anyhow::bail!(
@@ -40,6 +43,7 @@ impl std::fmt::Display for Toolchain {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
             Self::IrLLVM => write!(f, "ir-llvm"),
+            Self::ZkSolc => write!(f, "zksolc"),
             Self::Solc => write!(f, "solc"),
             Self::SolcLLVM => write!(f, "solc-llvm"),
         }
